+++ mobile-app/app/screens/AppNavigator/screens/Loans/LoansScreen.tsx
@@ -1,106 +1,57 @@
-import { useCallback, useEffect, useLayoutEffect, useState } from "react";
-import { TouchableOpacity, View } from "react-native";
-import { tailwind } from "@tailwind";
-import { ThemedIcon, ThemedView } from "@components/themed";
-import { Tabs } from "@components/Tabs";
-import {
-  SkeletonLoader,
-  SkeletonLoaderScreen,
-} from "@components/SkeletonLoader";
+import { useEffect, useRef, useState } from "react";
 import { batch, useSelector } from "react-redux";
+import { tailwind } from "@tailwind";
+import { translate } from "@translations";
 import { RootState } from "@store";
-import {
-  fetchLoanSchemes,
-  fetchLoanTokens,
-  fetchVaults,
-  loanTokensSelector,
-} from "@store/loans";
+import { fetchLoanSchemes, fetchLoanTokens, fetchVaults } from "@store/loans";
 import { useWhaleApiClient } from "@shared-contexts/WhaleContext";
 import { useWalletContext } from "@shared-contexts/WalletContext";
-import { StackScreenProps } from "@react-navigation/stack";
-import { HeaderSearchIcon } from "@components/HeaderSearchIcon";
-import { HeaderSearchInput } from "@components/HeaderSearchInput";
-import { debounce } from "lodash";
-import { LoanToken } from "@defichain/whale-api-client/dist/api/loan";
-import { useIsFocused } from "@react-navigation/native";
+import { useIsFocused, useScrollToTop } from "@react-navigation/native";
 import { useAppDispatch } from "@hooks/useAppDispatch";
-import { Vaults } from "@screens/AppNavigator/screens/Loans/components/Vaults";
-import { LoanParamList } from "./LoansNavigator";
+import { ThemedViewV2 } from "@components/themed";
+import { View } from "@components";
 import { LoanCards } from "./components/LoanCards";
-import { EmptyVault } from "./components/EmptyVault";
+import { Vaults } from "./components/Vaults";
+import { ButtonGroupV2 } from "../Dex/components/ButtonGroupV2";
 
 enum TabKey {
-  BrowseLoans = "BROWSE_LOANS",
+  Borrow = "BORROW",
   YourVaults = "YOUR_VAULTS",
 }
 
-type Props = StackScreenProps<LoanParamList, "LoansScreen">;
-
-export function LoansScreen({ navigation }: Props): JSX.Element {
+export function LoansScreen(): JSX.Element {
   const { address } = useWalletContext();
   const isFocused = useIsFocused();
   const blockCount = useSelector((state: RootState) => state.block.count);
-  const { vaults, hasFetchedVaultsData, hasFetchedLoansData } = useSelector(
-    (state: RootState) => state.loans
-  );
-  const loans = useSelector((state: RootState) =>
-    loanTokensSelector(state.loans)
-  );
-  const [activeTab, setActiveTab] = useState<string>(TabKey.YourVaults);
   const dispatch = useAppDispatch();
   const client = useWhaleApiClient();
-  const onPress = (tabId: string): void => {
-    if (tabId === TabKey.YourVaults) {
-      setShowSearchInput(false);
-    } else if (searchString !== "") {
-      setShowSearchInput(true);
-    } else {
-      // no-op: maintain search input state if no query
-    }
-    setActiveTab(tabId);
+
+  const [activeTab, setActiveTab] = useState<string>(TabKey.Borrow);
+
+  const onTabChange = (tabKey: TabKey): void => {
+    setActiveTab(tabKey);
   };
+
   const tabsList = [
     {
-      id: TabKey.YourVaults,
-      label: "Your vaults",
+      id: TabKey.Borrow,
+      label: translate("components/tabs", "Borrow"),
       disabled: false,
-      handleOnPress: onPress,
+      handleOnPress: () => onTabChange(TabKey.Borrow),
     },
     {
-      id: TabKey.BrowseLoans,
-      label: "Browse loan tokens",
+      id: TabKey.YourVaults,
+      label: translate("components/tabs", "Your vaults"),
       disabled: false,
-      handleOnPress: onPress,
+      handleOnPress: () => onTabChange(TabKey.YourVaults),
     },
   ];
 
-  // Search
-  const [filteredLoans, setFilteredLoans] = useState<LoanToken[]>(loans);
-  const [showSeachInput, setShowSearchInput] = useState(false);
-  const [searchString, setSearchString] = useState("");
-  const handleFilter = useCallback(
-    debounce((searchString: string) => {
-      setFilteredLoans(
-        loans.filter((loan) =>
-          loan.token.displaySymbol
-            .toLowerCase()
-            .includes(searchString.trim().toLowerCase())
-        )
-      );
-    }, 500),
-    [loans, hasFetchedLoansData]
-  );
-
-  useEffect(() => {
-    if (loans.length === 0) {
-      return;
-    }
-    handleFilter(searchString);
-  }, [searchString]);
-
-  useEffect(() => {
-    setFilteredLoans(loans);
-  }, [hasFetchedLoansData]);
+  /* useScrollToTop will not function in nested screens component, that's why we are passing the scroll view reference into LoadCards and Vaults component */
+  const borrowScrollRef = useRef(null);
+  const vaultScrollRef = useRef(null);
+  useScrollToTop(borrowScrollRef);
+  useScrollToTop(vaultScrollRef);
 
   useEffect(() => {
     if (isFocused) {
@@ -120,94 +71,39 @@ export function LoansScreen({ navigation }: Props): JSX.Element {
     dispatch(fetchLoanSchemes({ client }));
   }, []);
 
-  useLayoutEffect(() => {
-    navigation.setOptions({
-      headerRight: (): JSX.Element => {
-        if (activeTab === TabKey.BrowseLoans && vaults.length !== 0) {
-          return (
-            <HeaderSearchIcon
-              testID="header_loans_search"
-              onPress={() => setShowSearchInput(true)}
-            />
-          );
-        } else {
-          return (
-            <TouchableOpacity
-              // eslint-disable-next-line
-              onPress={() =>
-                navigation.navigate({
-                  name: "CreateVaultScreen",
-                  params: {},
-                  merge: true,
-                })
-              }
-              testID="create_vault_header_button"
-            >
-              <ThemedIcon
-                size={28}
-                style={tailwind("mr-2")}
-                light={tailwind("text-primary-500")}
-                dark={tailwind("text-primary-500")}
-                iconType="MaterialCommunityIcons"
-                name="plus"
-              />
-            </TouchableOpacity>
-          );
-        }
-      },
-    });
-  }, [navigation, activeTab, vaults]);
-
-  useEffect(() => {
-    if (showSeachInput) {
-      navigation.setOptions({
-        header: (): JSX.Element => (
-          <HeaderSearchInput
-            searchString={searchString}
-            onClearInput={() => setSearchString("")}
-            onChangeInput={(text: string) => setSearchString(text)}
-            onCancelPress={() => {
-              setSearchString("");
-              setShowSearchInput(false);
-            }}
-            placeholder="Search for loans"
-            testID="loans_search_input"
-          />
-        ),
-      });
-    } else {
-      navigation.setOptions({
-        header: undefined,
-      });
-    }
-  }, [showSeachInput, searchString]);
-
-  if (!hasFetchedVaultsData) {
-    return (
-      <View style={tailwind("mt-1")}>
-        <SkeletonLoader row={3} screen={SkeletonLoaderScreen.Vault} />
-      </View>
-    );
-  } else if (vaults?.length === 0) {
-    return <EmptyVault handleRefresh={() => {}} isLoading={false} />;
-  }
-
   return (
-    <ThemedView testID="loans_screen" style={tailwind("flex-1")}>
-      <Tabs
-        tabSections={tabsList}
-        testID="loans_tabs"
-        activeTabKey={activeTab}
-      />
-      {activeTab === TabKey.YourVaults && <Vaults />}
-      {activeTab === TabKey.BrowseLoans && !hasFetchedLoansData && (
-        <View style={tailwind("mt-1")}>
-          <SkeletonLoader row={6} screen={SkeletonLoaderScreen.Loan} />
+    <ThemedViewV2 testID="loans_screen" style={tailwind("flex-1")}>
+      <ThemedViewV2
+        light={tailwind("bg-mono-light-v2-00 border-mono-light-v2-100")}
+        dark={tailwind("bg-mono-dark-v2-00 border-mono-dark-v2-100")}
+        style={tailwind(
+          "flex flex-col items-center pt-4 rounded-b-2xl border-b"
+        )}
+      >
+        <View style={tailwind("w-full px-5")}>
+          <ButtonGroupV2
+            buttons={tabsList}
+            activeButtonGroupItem={activeTab}
+            testID="loans_tabs"
+            lightThemeStyle={tailwind("bg-transparent")}
+            darkThemeStyle={tailwind("bg-transparent")}
+          />
         </View>
-      )}
-      {activeTab === TabKey.BrowseLoans && hasFetchedLoansData && (
-        <LoanCards testID="loans_cards" loans={filteredLoans} />
-      )}
-    </ThemedView>
+      </ThemedViewV2>
+      <View
+        style={tailwind("flex-1", {
+          hidden: activeTab !== TabKey.YourVaults,
+        })}
+      >
+        <Vaults scrollRef={vaultScrollRef} />
+      </View>
+      <View
+        style={tailwind("flex-1", {
+          hidden: activeTab !== TabKey.Borrow,
+        })}
+      >
+        <LoanCards scrollRef={borrowScrollRef} testID="loan_screen" />
+      </View>
+    </ThemedViewV2>
   );
 }
