+++ mobile-app/app/screens/AppNavigator/screens/Loans/components/VaultCard.tsx
@@ -1,44 +1,59 @@
+import { useMemo } from "react";
+import { View } from "react-native";
+import { useSelector } from "react-redux";
 import BigNumber from "bignumber.js";
-import {
-  ThemedIcon,
-  ThemedText,
-  ThemedTouchableOpacity,
-  ThemedView,
-} from "@components/themed";
-import { tailwind } from "@tailwind";
-import { View } from "@components";
-import { translate } from "@translations";
-import { TokenIconGroup } from "@components/TokenIconGroup";
-import { IconButton } from "@components/IconButton";
 import { NavigationProp, useNavigation } from "@react-navigation/native";
-import { LoanParamList } from "@screens/AppNavigator/screens/Loans/LoansNavigator";
-import { LoanVault } from "@store/loans";
-import { LoanVaultActive } from "@defichain/whale-api-client/dist/api/loan";
-import { VaultInfo } from "@screens/AppNavigator/screens/Loans/components/VaultInfo";
-import { TouchableOpacity } from "react-native";
-import { openURL } from "@api/linking";
-import { useDeFiScanContext } from "@shared-contexts/DeFiScanContext";
-import { VaultSectionTextRow } from "@screens/AppNavigator/screens/Loans/components/VaultSectionTextRow";
+import { tailwind } from "@tailwind";
+import { translate } from "@translations";
 import {
-  useVaultStatus,
-  VaultStatusTag,
-} from "@screens/AppNavigator/screens/Loans/components/VaultStatusTag";
-import { useNextCollateralizationRatio } from "@screens/AppNavigator/screens/Loans/hooks/NextCollateralizationRatio";
-import { CollateralizationRatioDisplay } from "@screens/AppNavigator/screens/Loans/components/CollateralizationRatioDisplay";
-import { TabKey } from "@screens/AppNavigator/screens/Loans/VaultDetail/components/VaultDetailTabSection";
-import { useLoanOperations } from "@screens/AppNavigator/screens/Loans/hooks/LoanOperations";
+  LoanToken,
+  LoanVaultActive,
+} from "@defichain/whale-api-client/dist/api/loan";
+import { RootState } from "@store";
+import { loanTokensSelector, LoanVault } from "@store/loans";
+import { LoanParamList } from "@screens/AppNavigator/screens/Loans/LoansNavigator";
+import { useVaultStatus } from "@screens/AppNavigator/screens/Loans/components/VaultStatusTag";
 import { VaultStatus } from "@screens/AppNavigator/screens/Loans/VaultStatusTypes";
-import { getPrecisedTokenValue } from "@screens/AppNavigator/screens/Auctions/helpers/precision-token-value";
+import { getPrecisedCurrencyValue } from "@screens/AppNavigator/screens/Auctions/helpers/precision-token-value";
+import { VaultCardStatus } from "@screens/AppNavigator/screens/Loans/components/VaultCardStatus";
+import { ThemedTouchableOpacityV2, ThemedViewV2 } from "@components/themed";
+import { BottomSheetNavScreen } from "@components/BottomSheetWithNavV2";
+import { TokenIconGroupV2 } from "@components/TokenIconGroupV2";
+import { BottomSheetTokenListHeader } from "@components/BottomSheetTokenListHeader";
+import {
+  BottomSheetTokenList,
+  TokenType,
+} from "@components/BottomSheetTokenList";
+import { VaultSectionTextRowV2 } from "./VaultSectionTextRowV2";
+import { VaultBanner } from "./VaultBanner";
+import { useCollateralTokenList } from "../hooks/CollateralTokenList";
 
 export interface VaultCardProps extends React.ComponentProps<any> {
   vault: LoanVault;
   testID: string;
+  dismissModal: () => void;
+  expandModal: () => void;
+  setBottomSheetScreen: React.Dispatch<
+    React.SetStateAction<BottomSheetNavScreen[]>
+  >;
+  setSnapPoints: React.Dispatch<
+    React.SetStateAction<{
+      ios: string[];
+      android: string[];
+    }>
+  >;
+  onBottomSheetLoansTokensListSelect: ({
+    onPress,
+    loanTokens,
+  }: {
+    onPress: (item: LoanToken) => void;
+    loanTokens: LoanToken[];
+  }) => void;
 }
 
 export function VaultCard(props: VaultCardProps): JSX.Element {
   const navigation = useNavigation<NavigationProp<LoanParamList>>();
   const vault = props.vault as LoanVaultActive;
-  const { getVaultsUrl } = useDeFiScanContext();
   const vaultState = useVaultStatus(
     vault.state,
     new BigNumber(vault.informativeRatio),
@@ -46,212 +61,170 @@ export function VaultCard(props: VaultCardProps): JSX.Element {
     new BigNumber(vault.loanValue),
     new BigNumber(vault.collateralValue)
   );
-  const nextCollateralizationRatio = useNextCollateralizationRatio(
-    vault.collateralAmounts,
-    vault.loanAmounts
+
+  const { collateralTokens } = useCollateralTokenList();
+  const loanTokens = useSelector((state: RootState) =>
+    loanTokensSelector(state.loans)
   );
-  const canUseOperations = useLoanOperations(vault?.state);
+
+  const onSelectLoanToken = (item: LoanToken) => {
+    props.dismissModal();
+    navigation.navigate({
+      name: "BorrowLoanTokenScreen",
+      params: {
+        vault: vault,
+        loanToken: item,
+      },
+      merge: true,
+    });
+  };
+
+  const onBottomSheetLoansTokensListSelect = (): void => {
+    props.onBottomSheetLoansTokensListSelect({
+      onPress: onSelectLoanToken,
+      loanTokens: loanTokens,
+    });
+  };
+
   const onCardPress = (): void => {
     navigation.navigate("VaultDetailScreen", {
       vaultId: vault.vaultId,
     });
   };
+
+  const onAddCollateralPress = (): void => {
+    props.setSnapPoints({ ios: ["70%"], android: ["70%"] });
+    props.setBottomSheetScreen([
+      {
+        stackScreenName: "TokenList",
+        component: BottomSheetTokenList({
+          tokenType: TokenType.CollateralItem,
+          vault: vault,
+          onTokenPress: async (item) => {
+            navigation.navigate({
+              name: "AddOrRemoveCollateralScreen",
+              params: {
+                vault,
+                collateralItem: item,
+                collateralTokens,
+                isAdd: true,
+              },
+            });
+            props.dismissModal();
+          },
+        }),
+        option: {
+          headerTitle: "",
+          headerBackTitleVisible: false,
+          headerStyle: tailwind("rounded-t-xl-v2 border-b-0"),
+          header: () => (
+            <BottomSheetTokenListHeader
+              headerLabel={translate(
+                "screens/EditCollateralScreen",
+                "Select Collateral"
+              )}
+              onCloseButtonPress={props.dismissModal}
+            />
+          ),
+        },
+      },
+    ]);
+    props.expandModal();
+  };
+
+  const { isVaultEmpty, isVaultLiquidated } = useMemo(() => {
+    return {
+      isVaultEmpty: vaultState.status === VaultStatus.Empty,
+      isVaultLiquidated: vaultState.status === VaultStatus.Liquidated,
+    };
+  }, [vaultState.status]);
+
+  const vaultBanner = useMemo(() => {
+    let text = "";
+    let type = "";
+    let buttonLabel = "";
+    switch (vaultState.status) {
+      case VaultStatus.Liquidated:
+        text = "Liquidated vault";
+        type = VaultStatus.Liquidated;
+        break;
+      case VaultStatus.Empty:
+        text = "Add collateral to borrow";
+        type = VaultStatus.Empty;
+        buttonLabel = "Add Collateral";
+        break;
+      default:
+        break;
+    }
+    return { buttonLabel, text, type };
+  }, [vaultState.status]);
+
   return (
-    <ThemedView
-      light={tailwind("bg-white border-gray-200")}
-      dark={tailwind("bg-gray-800 border-gray-700")}
-      style={tailwind("rounded mb-2 border p-4")}
-    >
-      <ThemedTouchableOpacity
-        testID={props.testID}
-        onPress={onCardPress}
-        style={tailwind("mb-2")}
-        light={tailwind("border-b-0")}
-        dark={tailwind("border-b-0")}
-      >
-        <View style={tailwind("flex flex-row justify-between mb-4")}>
-          <View style={tailwind("flex flex-row flex-1 items-center")}>
-            <View style={tailwind("flex flex-col flex-1")}>
-              <View
-                style={tailwind("flex flex-row justify-between items-center")}
-              >
-                <View style={tailwind("flex-col")}>
-                  <View style={tailwind("flex-row items-center")}>
-                    <ThemedText
-                      light={tailwind("text-gray-400")}
-                      dark={tailwind("text-gray-500")}
-                      style={tailwind("text-xs mr-1.5")}
-                    >
-                      {translate("screens/VaultDetailScreen", "Vault ID")}
-                    </ThemedText>
-                    <VaultStatusTag
-                      status={vaultState.status}
-                      testID={`${props.testID}_status`}
-                    />
-                  </View>
-                  <TouchableOpacity
-                    style={tailwind("flex flex-row mb-0.5 items-center")}
-                    onPress={async () =>
-                      await openURL(getVaultsUrl(vault.vaultId))
-                    }
-                  >
-                    <ThemedText
-                      testID={`${props.testID}_vault_id`}
-                      style={tailwind("font-semibold w-56 flex-shrink mr-0.5")}
-                      numberOfLines={1}
-                      ellipsizeMode="middle"
-                    >
-                      {vault.vaultId}
-                    </ThemedText>
-                    <ThemedIcon
-                      dark={tailwind("text-darkprimary-500")}
-                      iconType="MaterialIcons"
-                      light={tailwind("text-primary-500")}
-                      name="open-in-new"
-                      size={18}
-                    />
-                  </TouchableOpacity>
-                </View>
-                {canUseOperations && (
-                  <ThemedIcon
-                    dark={tailwind("text-gray-200")}
-                    iconType="MaterialIcons"
-                    light={tailwind("text-black")}
-                    name="chevron-right"
-                    size={30}
-                  />
-                )}
-              </View>
-              <View style={tailwind("flex flex-row")}>
-                <ThemedText
-                  light={tailwind("text-gray-500")}
-                  dark={tailwind("text-gray-300")}
-                  style={tailwind("text-xs mr-1")}
-                >
-                  {translate("components/VaultCard", "Collateral:")}
-                </ThemedText>
-                {(vault.collateralAmounts?.length === 0 ||
-                  vault.collateralAmounts === undefined) && (
-                  <ThemedText
-                    light={tailwind("text-gray-500")}
-                    dark={tailwind("text-gray-300")}
-                    style={tailwind("text-xs ml-1")}
-                    testID={`${props.testID}_collateral_none`}
-                  >
-                    {translate("components/VaultCard", "None")}
-                  </ThemedText>
-                )}
-                <TokenIconGroup
+    <View style={tailwind("mb-2")}>
+      {isVaultEmpty || isVaultLiquidated ? (
+        <VaultBanner
+          buttonLabel={vaultBanner.buttonLabel}
+          description={vaultBanner.text}
+          vaultId={vault.vaultId}
+          onButtonPress={onAddCollateralPress}
+          vaultType={vaultBanner.type}
+          onCardPress={onCardPress}
+          testID={`${props.testID}_${vaultBanner.type}`}
+          vault={vault}
+        />
+      ) : (
+        <ThemedTouchableOpacityV2
+          onPress={onCardPress}
+          style={tailwind("border-0")}
+        >
+          <View style={tailwind("flex-col")}>
+            <ThemedViewV2
+              dark={tailwind("bg-mono-dark-v2-00")}
+              light={tailwind("bg-mono-light-v2-00")}
+              style={tailwind(
+                "flex-row p-5 rounded-lg-v2 border-0 items-center justify-between"
+              )}
+              testID={props.testID}
+            >
+              <View style={tailwind("flex-1 flex-col pr-2")}>
+                <TokenIconGroupV2
                   testID={`${props.testID}_collateral_token_group`}
                   symbols={vault.collateralAmounts?.map(
                     (collateral) => collateral.displaySymbol
                   )}
-                  maxIconToDisplay={3}
+                  maxIconToDisplay={6}
+                  size={24}
+                />
+                <VaultSectionTextRowV2
+                  testID={`${props.testID}_loan_available`}
+                  prefix="$"
+                  value={getPrecisedCurrencyValue(vault.loanValue) ?? "-"}
+                  lhs={translate("components/VaultCard", "Total loans")}
+                  isOraclePrice
+                  customContainerStyle="mt-3"
+                />
+                <VaultSectionTextRowV2
+                  testID={`${props.testID}_total_collateral`}
+                  prefix="$"
+                  value={getPrecisedCurrencyValue(vault.collateralValue)}
+                  lhs={translate("components/VaultCard", "Total collateral")}
+                  isOraclePrice
+                  customContainerStyle="mt-2"
                 />
               </View>
-            </View>
-          </View>
-        </View>
-        {![
-          VaultStatus.Empty,
-          VaultStatus.Ready,
-          VaultStatus.Unknown,
-          VaultStatus.Liquidated,
-        ].includes(vaultState.status) && (
-          <CollateralizationRatioDisplay
-            testID={props.testID}
-            collateralizationRatio={vault.informativeRatio}
-            minCollateralizationRatio={vault.loanScheme.minColRatio}
-            totalLoanAmount={vault.loanValue}
-            nextCollateralizationRatio={nextCollateralizationRatio?.toFixed(8)}
-          />
-        )}
-        <View style={tailwind("flex flex-row flex-wrap -mb-2")}>
-          {vault.loanAmounts?.length > 0 && (
-            <VaultInfo
-              label="Active loans"
-              testID={props.testID}
-              tokens={vault.loanAmounts?.map((loan) => loan.displaySymbol)}
-            />
-          )}
-          <VaultSectionTextRow
-            testID={`${props.testID}_total_loan`}
-            prefix={VaultStatus.Liquidated === vaultState.status ? "" : "$"}
-            value={
-              VaultStatus.Liquidated === vaultState.status
-                ? "-"
-                : getPrecisedTokenValue(vault.loanValue) ?? "-"
-            }
-            lhs={translate("components/VaultCard", "Total loans (USD)")}
-            isOraclePrice
-          />
-          <VaultSectionTextRow
-            testID={`${props.testID}_total_collateral`}
-            prefix={VaultStatus.Liquidated === vaultState.status ? "" : "$"}
-            value={
-              VaultStatus.Liquidated === vaultState.status
-                ? "-"
-                : getPrecisedTokenValue(vault.collateralValue)
-            }
-            lhs={translate("components/VaultCard", "Total collateral (USD)")}
-            isOraclePrice
-          />
-        </View>
-      </ThemedTouchableOpacity>
-      <VaultActionButton
-        vault={vault}
-        canUseOperation={canUseOperations}
-        testID={props.testID}
-      />
-    </ThemedView>
-  );
-}
 
-function VaultActionButton({
-  vault,
-  canUseOperation,
-  testID,
-}: {
-  vault: LoanVaultActive;
-  canUseOperation: boolean;
-  testID: string;
-}): JSX.Element | null {
-  const navigation = useNavigation<NavigationProp<LoanParamList>>();
-  return (
-    <ThemedView
-      light={tailwind("border-gray-200")}
-      dark={tailwind("border-gray-700")}
-      style={tailwind("flex flex-row mt-4 items-center flex-wrap -mb-2")}
-    >
-      {new BigNumber(vault.collateralValue).gt(0) && (
-        <IconButton
-          testID={`${testID}_manage_loans_button`}
-          disabled={!canUseOperation}
-          iconLabel={translate("components/VaultCard", "MANAGE LOANS")}
-          style={tailwind("mr-2 mb-2 items-center")}
-          onPress={() => {
-            navigation.navigate("VaultDetailScreen", {
-              vaultId: vault.vaultId,
-              tab: TabKey.Loans,
-            });
-          }}
-        />
+              <VaultCardStatus
+                vault={vault}
+                vaultStatus={vaultState.status}
+                colRatio={vault.informativeRatio}
+                minColRatio={vault.loanScheme.minColRatio}
+                onButtonPressed={onBottomSheetLoansTokensListSelect}
+                testID={props.testID}
+              />
+            </ThemedViewV2>
+          </View>
+        </ThemedTouchableOpacityV2>
       )}
-      <IconButton
-        testID={`${testID}_edit_collaterals_button`}
-        disabled={!canUseOperation}
-        iconLabel={translate("components/VaultCard", "EDIT COLLATERAL")}
-        style={tailwind("mr-2 mb-2 items-center")}
-        onPress={() => {
-          navigation.navigate({
-            name: "EditCollateralScreen",
-            params: {
-              vaultId: vault.vaultId,
-            },
-            merge: true,
-          });
-        }}
-      />
-    </ThemedView>
+    </View>
   );
 }
