--- mobile-app/app/screens/AppNavigator/screens/Loans/screens/ChooseLoanTokenScreenV2.tsx
@@ -1,17 +1,0 @@
-import { ThemedView } from "@components/themed";
-import { StackScreenProps } from "@react-navigation/stack";
-import { tailwind } from "@tailwind";
-import { LoanParamList } from "../LoansNavigator";
-import { LoanCardsV2 } from "../components/LoanCardsV2";
-
-type Props = StackScreenProps<LoanParamList, "ChooseLoanTokenScreen">;
-
-export function ChooseLoanTokenScreenV2({ route }: Props): JSX.Element {
-  const { vaultId } = route.params;
-
-  return (
-    <ThemedView style={tailwind("flex-1")}>
-      <LoanCardsV2 testID="loans_cards" vaultId={vaultId} />
-    </ThemedView>
-  );
-}
